requirements.txt
streamlit>=1.32.0
cloudinary>=1.36.0
python-dotenv>=1.0.1
Pillow>=10.2.0
openai>=1.12.0
pandas>=2.2.0 